+++ core/utils/lume_config.ts
@@ -12,6 +12,7 @@ export const pluginNames = [
   "eta",
   "extract_date",
   "extract_order",
+  "git_date",
   "jsx",
   "json_ld",
   "reading_info",