+++ tests/git_date.test.ts
@@ -1,0 +1,26 @@
+import { assertSiteSnapshot, build, getSite } from "./utils.ts";
+import gitDate from "../plugins/git_date.ts";
+
+Deno.test("extract_date plugin", async (t) => {
+  const site = getSite({
+    src: "git_date",
+  });
+
+  site.use(gitDate());
+
+  await build(site);
+  await assertSiteSnapshot(t, site);
+});
+
+Deno.test("extract_date plugin (override)", async (t) => {
+  const site = getSite({
+    src: "git_date",
+  });
+
+  site.use(gitDate({
+    override: true,
+  }));
+
+  await build(site);
+  await assertSiteSnapshot(t, site);
+});